requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
redis==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
